Move some resolvers to server-side-only
