+++ packages/lesswrong/lib/collections/petrovDayLaunchs/index.ts
@@ -1,4 +1,3 @@
 export * from './collection';
 import './schema';
 import './fragments';
-import './mutations';
